+++ config/OptionsChainLoader.js
@@ -1,9 +1,0 @@
-module.exports = ({ config, resolve }) => {
-  const baseRule = config.module.rule('js').test(/.js|.tsx?$/);
-  const normalRule = baseRule.oneOf('normal');
-  return () => {
-    normalRule
-      .use('options-chain')
-      .loader(resolve('options-chain-loader'))
-  }
-}