+++ CHANGELOG.md
@@ -3,18 +3,19 @@
 ### 🌟 新功能
 
 * **课题2：** 完成基本构建，js、css打包进html，实现dev跟build环境 ([71a6a19](https://github.com/luoxue-victor/learn_webpack/commit/71a6a19))
+* **新增loader配置：** 增加ts、css、less、sass、postcss、babel配置 ([53616f8](https://github.com/luoxue-victor/learn_webpack/commit/53616f8))
 * **修改了配置文件：** i feat a config ([382c60c](https://github.com/luoxue-victor/learn_webpack/commit/382c60c))
 * **all：** 初始化一个项目 ([336cc13](https://github.com/luoxue-victor/learn_webpack/commit/336cc13))
 * **asdasdasdas：** 你爱哈的 ([ea12f9c](https://github.com/luoxue-victor/learn_webpack/commit/ea12f9c))
 * **config：** add config  option ([83c9f1e](https://github.com/luoxue-victor/learn_webpack/commit/83c9f1e))
 * **daasd：** asdasd ([0b6ee85](https://github.com/luoxue-victor/learn_webpack/commit/0b6ee85))
-* **feat：** aaaaa ([73abf42](https://github.com/luoxue-victor/learn_webpack/commit/73abf42))
+* **feat：** asdada ([64090c7](https://github.com/luoxue-victor/learn_webpack/commit/64090c7))
 * **init：** 项目初始化 ([d7835fb](https://github.com/luoxue-victor/learn_webpack/commit/d7835fb))
 * **webpack：** webpack配置 ([9a1d29c](https://github.com/luoxue-victor/learn_webpack/commit/9a1d29c))
 * 你好 ([6e534da](https://github.com/luoxue-victor/learn_webpack/commit/6e534da))
 * asasdasasd ([8f9515f](https://github.com/luoxue-victor/learn_webpack/commit/8f9515f))
 * asd ([d278787](https://github.com/luoxue-victor/learn_webpack/commit/d278787))
-* **feat：** asdada ([64090c7](https://github.com/luoxue-victor/learn_webpack/commit/64090c7))
+* **feat：** aaaaa ([73abf42](https://github.com/luoxue-victor/learn_webpack/commit/73abf42))
 
 
 ### 🐛 Bug 修复